\name{spmeshed-package}
\alias{spmeshed-package}
\alias{spmeshed}
\docType{package}
\title{
\packageTitle{spmeshed}
}
\description{
\packageDescription{spmeshed}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spmeshed}
\packageIndices{spmeshed}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{spmeshed}

Maintainer: \packageMaintainer{spmeshed}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
